unclutter tests
